pdf_extractor/requirements.txt
PyQt6
pymupdf
pytesseract
Pillow
pytest
